+++ TODO.md
@@ -1,4 +1,9 @@
-- [x] Update generateMetadata in src/app/(pages)/events/[id]/page.tsx to await params
-- [x] Update PageProps type in page.tsx to params: Promise<{ id: string }>
-- [x] Update the EventPage component to await params
-- [x] Run Next.js build to verify the fix
+# TODO: Add Header and Footer to VolunteerForm
+
+- [x] Import Navbar and Footer components in VolunteerForm.tsx
+- [x] Change the root div from fixed modal overlay to full-page layout
+- [x] Add Navbar component at the top of the layout
+- [x] Adjust form content positioning to center it in the page
+- [x] Add Footer component at the bottom of the layout
+- [x] Remove or reposition Marque component if it conflicts with Footer
+- [x] Test the layout to ensure form functionality and UI remain unchanged
