+++ src/app/(pages)/joinus/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import SuccessModal from "@/app/components/SuccessModal";
-import VolunteerForm from "@/app/(pages)/form/page";
+import VolunteerForm from "@/components/VolunteerForm";
 import Image from "next/image";
 import hero_img1 from "../../../../public/hero-img1.svg";
 import hero_img2 from "../../../../public/hero-img2.svg";
@@ -65,7 +65,7 @@ export default function JoinUsPage() {
       {/* Volunteer Form Modal */}
       {isModalOpen && (
         <VolunteerForm
-          setOpenModal={setIsModalOpen}
+          // setOpenModal={setIsModalOpen}
           onSuccess={() => {
             setIsModalOpen(false);
             setIsSuccessModalOpen(true);
